requirements.txt
flask
flask-cors
bcrypt
pyjwt
gunicorn
psycopg2-binary
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
flask-jwt-extended
flask-bcrypt